+++ library/docs/changelog.md
@@ -8,7 +8,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Groovy DSL for adding feature flags via Gradle plugin. This is equivalent to `feature("SomeFeatureFlag")` function.
-
     ```groovy
     laboratory {
       SomeFeatureFlag {
@@ -17,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       }
     }
     ```
+- `options` extension to `Class<Feature<T>>` that returns all available feature flag options.
 
 ### Changed
 - `excludeProjects` plugin functions are now called `projectFilter` and the condition is reversed. Previously they removed projects that matched a condition. Now they allow projects that match it.